feat(solver): the compiled walk is bit-identical, and selectable
The last five entries were NEP 50. `tree_traversal` accumulates the average
strategy as `strategy_sum[slot] += probability * weight`, where `probability`
comes from `.tolist()` and is therefore a PYTHON float -- and under NumPy 2 a
Python scalar is weak, so numpy performs that entire add in float32. Numba
types the same expression as float64 and rounds on the store, which lands a
different float about once in 150,000 accumulations. Narrowing the addend to
float32 closes it.

Worth keeping: the four earlier hypotheses were all wrong and all disproved by
measurement (numba's float32 accumulate on 200k values, the strategy vector,
FMA contraction, explicit widening). What finally located it was arithmetic on
the numbers themselves -- all five slots held the SAME value, so the strategy
was uniform and the addends identical, which left only the accumulate.

  BIT-IDENTICAL over 250 iterations: regrets, strategy_sum, reach_counts,
  cumulative_utility, visited, every per-iteration utility, and both random
  streams left in the same place.

Selected by `solver.traversal="compiled"`, default still "tree". Overridden at
`train_iteration` rather than `_cfr_external_sampling` because the kernel owns
the whole iteration -- it deals, walks, and advances both generators in one
crossing. The context is built on first use, since it needs the abstraction's
raw arrays and a solver over a stub bucketer has none.

Not yet the default: this has run in-process only. The 16-worker shared-memory
path has never executed it.

+++ src/engine/solver/mccfr/compiled_walk.py
@@ -430,7 +430,13 @@ def walk(
         # Widen, add, round ONCE — numpy's `float32[i] += float64` computes in
         # float64 and rounds on the store, and narrowing the addend first
         # instead lands a different float on about one accumulation in 150,000.
-        strategy_sum[start + i] += strategy[i] * strategy_weight
+        # float32, NOT float64. `tree_traversal` does
+        # `strategy_sum[slot] += probability * weight` where `probability` came
+        # from `.tolist()` and is a PYTHON float -- and under NEP 50 a Python
+        # scalar is weak, so numpy computes that add entirely in float32.
+        # Numba types the same expression as float64 and rounds on the store,
+        # which is a different result about once in 150,000 accumulations.
+        strategy_sum[start + i] += np.float32(strategy[i] * strategy_weight)
     applied += 1
 
     draw, sample_index = random_sample(sample_state, sample_index)
